refactor(shared): extract LLM utilities from deep-search to shared/llm-utils
- Move llmComplete, llmCompleteWithTool, extractJSON and internal helpers
  (chatCompletion, resolveConfig, types) to src/shared/llm-utils.ts
- Update memory/ consumers (topic-consolidator, knowledge-extractor) to
  import from shared/ — eliminates reverse dependency on tools/workflow/
- Update deep-search internals (engine, quality-gate, tests) to import
  from shared/llm-utils directly
- Mark tools.md §8.5 as Done
