+++ src/memory/topic-consolidator.ts
@@ -13,7 +13,7 @@
 
 import fs from "node:fs";
 import path from "node:path";
-import { llmCompleteWithTool } from "../tools/workflow/deep-search/sub-agent.js";
+import { llmCompleteWithTool } from "../shared/llm-utils.js";
 
 // ---------------------------------------------------------------------------
 // Thresholds
