+++ src/tools/workflow/deep-search/engine.ts
@@ -1,4 +1,5 @@
-import { runSubAgent, llmComplete, llmCompleteWithTool, extractJSON, type SubAgentOptions, type ProgressCallback, type ProgressEvent } from "./sub-agent.js";
+import { runSubAgent, type SubAgentOptions, type ProgressCallback, type ProgressEvent } from "./sub-agent.js";
+import { llmComplete, llmCompleteWithTool, extractJSON } from "../../../shared/llm-utils.js";
 import {
   contextGatheringPrompt,
   hypothesisGenerationPrompt,
